[CHORE] UiState 파일로 분리
